requirements.txt
<<<<<<< HEAD
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2021.10.26=haa95532_4
certifi=2021.10.8=py39haa95532_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
conda=4.11.0=py39haa95532_0
conda-package-handling=1.7.3=py39h8cc25b3_1
console_shortcut=0.1.1=4
contourpy=1.2.1=pypi_0
cryptography=36.0.0=py39h21b164f_0
cycler=0.12.1=pypi_0
fonttools=4.52.4=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.34.1=pypi_0
importlib-resources=6.4.0=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
matplotlib=3.9.0=pypi_0
menuinst=1.4.16=py39h2bbff1b_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=1.1.1m=h2bbff1b_0
packaging=24.0=pypi_0
pillow=10.3.0=pypi_0
pip=21.2.4=py39haa95532_0
powershell_shortcut=0.0.1=3
pycosat=0.6.3=py39h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.5=h6244533_3
python-dateutil=2.9.0.post0=pypi_0
pywin32=302=py39h827c3e9_1
requests=2.27.1=pyhd3eb1b0_0
ruamel_yaml=0.15.100=py39h2bbff1b_0
scikit-image=0.22.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=58.0.4=py39haa95532_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=h2bbff1b_0
tifffile=2024.5.22=pypi_0
tqdm=4.62.3=pyhd3eb1b0_1
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
yaml=0.2.5=he774522_0
zipp=3.19.0=pypi_0
=======
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.4
imageio==2.34.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
tifffile==2024.5.22
tqdm==4.66.4
>>>>>>> a0923a6e61de6fadb9e0b9db5bf1f79a3c201609
